Refactor: 같은 페이지 type들 한 파일로 정리, 경로 수정
