+++ src/containers/main/HotInterest.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link'
 import Data from '@/dummydata/liveBidding.json'
-import { LiveAndHotType } from '@/types/LiveAndHotType'
+import { LiveAndHotType } from '@/types/mainType'
 import LiveAndHotItem from './LiveAndHotItem'
 
 export default function HotInterest() {
