+++ src/containers/main/TagItem.tsx
@@ -1,4 +1,4 @@
-import { Tag } from '@/types/SoonAndHitsType'
+import { Tag } from '@/types/mainType'
 
 export default function TagItem({ tag }: { tag: Tag }) {
   return <p className="pr-[3px] text-[14px] text-[#808080]">#{tag.tagName}</p>
