+++ src/containers/search/RecentSearch.tsx
@@ -3,7 +3,7 @@
 // eslint-disable-next-line import/no-cycle
 
 import Remove from '@/public/svgs/icon/remove.svg'
-import { KeywordType } from '@/types/KeywordType'
+import { KeywordType } from '@/types/headerType'
 
 export default function RecentSearch({
   keywords,
