--- src/types/BannerType.ts
@@ -1,4 +1,0 @@
-export interface BannerType {
-  id: number
-  url: string
-}
