--- src/types/KeywordType.ts
@@ -1,4 +1,0 @@
-export interface KeywordType {
-  id: number
-  text: string
-}
