--- src/types/LiveAndHotType.ts
@@ -1,11 +1,0 @@
-export interface LiveAndHotType {
-  goodsCode: string
-  thumbnail: string
-  goodsName: string
-  minPrice: number
-  closedAt: string
-  viewsCount: number
-  wishCount: number
-  biddingCount: number
-  tradingStatus: number
-}
