+++ src/types/headerType.ts
@@ -1,3 +1,8 @@
+export interface KeywordType {
+  id: number
+  text: string
+}
+
 export interface NotificationType {
   id: number
   code: string
